[ADD] balances by partner
