+++ gec_account/views/account_move_views.xml
@@ -6,13 +6,17 @@
         <field name="inherit_id" ref="account.view_move_form"/>
         <field name="arch" type="xml">
 
-            <xpath expr="//field[@name='line_ids']/tree//field[@name='name']" position="after">
-                <field name="test_field"/>
-                <widget name="customerBalances" width="20px"/>
+            <xpath expr="//field[@name='line_ids']/tree//field[@name='partner_id']" position="after">
+                <widget name="customer_balances" width="20px"/>
+                <field name="virtual_debe" invisible="1"/>
+                <field name="virtual_haber" invisible="1"/>
+                <field name="virtual_balance" invisible="1"/>
             </xpath>
-            <xpath expr="//field[@name='line_ids']/form//field[@name='name']" position="after">
-                <field name="test_field"/>
-            </xpath>
+
+
+<!--            <xpath expr="//field[@name='line_ids']/form//field[@name='name']" position="after">-->
+<!--                <field name="test_field"/>-->
+<!--            </xpath>-->
         </field>
     </record>
